requirements.txt
Django==5.2.13
djangorestframework==3.17.1
django-taggit==6.1.0
Pillow==12.2.0
markdown==3.10.2
django-crispy-forms==2.6
crispy-tailwind==1.0.3
python-decouple==3.8
whitenoise==6.12.0
